refactor: simplify discovery routes and extract handlers into separate files
